fix:feed Response 수정 및 ImageTag 좌표 Double로 수정

+++ server/src/main/java/com/mainproject/server/imagetag/dto/ImageTagResponseDto.java
@@ -13,7 +13,7 @@ public class ImageTagResponseDto {
     private String productName;
     private String productPrice;
     private String productInfo;
-    private Long positionX;
-    private Long positionY;
+    private Double positionX;
+    private Double positionY;
 
 }
